Refactor Image\Type into CoreType(s) which implement new TypeInterface. Type class is only a singleton now. Additional types can now be registered as well.
